==> production/02-analisis.Rmd <==
---
title: "Analysis"
date: "`r format(Sys.time(), '%A %d %B %Y %X')`"
output:
  html_document:
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: no
      number_sections: yes
    code_folding: hide
    number_sections: yes
editor_options:
  chunk_output_type: console
---
# Setup

```{r setup}
knitr::opts_chunk$set(
	echo = FALSE,
	results = "asis",
	message = FALSE,
	warning = FALSE, 
	fig.height = 14, 
	fig.width = 14, 
	cache = FALSE
)
options(scipen=9999) # desactivar notacion cientifica
remove(list = ls()) #limpieza del entorno de trabajo
```

```{css, echo=FALSE}
# /* expandir el espacio del documento*/
body .main-container {
      max-width: 1600px !important;
      width: 1600px !important;
    }
#/* expandir width de la TOC*/
div.tocify {
    width: 22% !important;
    max-width: 331px !important;
    max-height: 102% !important;
}
```

# A. Frecuencia: asiste a marchar

## Descriptivos

```{r tab-desc, results="asis", warning=FALSE}
if (!require("pacman")) install.packages("pacman") # instalar pacman
                            # cargar librerias
pacman::p_load(dplyr,       # Manipulacion de datos
               knitr,
               kableExtra,
               summarytools,
               ggplot2,
               sjPlot,
               tidyverse,
               ggalluvial,
               survey,
               shadowtext,
               srvyr,
               interactions
               )
load(file = here::here("input/data-proc/df_study1_long_t7.RData"))

# Css
st_css()
st_options(lang = "en",
           footnote = NA,
           bootstrap.css = F,
           custom.css = here::here("input/css/dfsummary.css"),
           dfSummary.custom.1 = NA
           )  

# Df variables nivel 1
# df<- dfSummary(df_study1_long %>% select(in_simpat:edad),
#                plain.ascii = FALSE,
#                style = "grid",
#                tmp.img.dir = "/tmp",
#                graph.magnif = TRUE,
#                headings = F,  # encabezado
#                varnumbers = F, # num variable
#                labels.col = T, # etiquetas
#                na.col = F,    # missing
#                graph.col = T, # plot
#                valid.col = T, # n valido
#                col.widths = c(700,10,10,10,10)
#                )
# df$Variable <- NULL # delete variable column
# view(df,
#      file = here::here("output/tables/desc01.html")
#      ) # Ver tabla en un archivo HTML
# webshot::webshot(url =here::here("output/tables/desc01.html"),
#         file =here::here("output/tables/desc01.png")) # Remover archivo HTML
# # print(df, method = "render")

# Df variables nivel 1 _________________________________________________________
df_study1 <- df_study1_long_t7 %>% dplyr::select(idencuesta,ola,region_cod,comuna,
         sexo,
         pos_id,
         edad,
         educ,
         quintil1,
         ess,
         barrio_ideal,
         barrio_integrado,
         barrio_identifico,
         barrio_partedemi,
         barrio_amigos,
         barrio_sociable,
         barrio_cordial,
         barrio_colaboradora,
         satisfaccion_proximidadtrabajo,
         satisfaccion_proximidadcolegios,
         satisfaccion_proximidadcomercio,
         marchar,
         part_com,
         satisfaccion_seguridad,
         jjvv_prom,
         org_caridad_prom,
         org_religiosa_prom,
         org_deportiva_prom
         )

df_study1$ola <- factor(df_study1$ola,
         labels = c("2016","2017","2018","2019", "2021", "2022", "2023"))
df_study1$ola <- 
sjlabelled::set_label(x = df_study1$ola,
                      label = "Ola")

df_study1$sexo <- factor(df_study1$sexo,
         labels = c("Hombre", "Mujer"))
df_study1$sexo <- 
sjlabelled::set_label(x = df_study1$sexo,
                      label = "Sexo")


df_study1$marchar <- 
sjlabelled::set_label(x = df_study1$marchar, 
                      label = "Frecuencia: asiste a marchar")

#df_study1$part_com <- 
#sjlabelled::set_label(x = df_study1$part_com, 
#                      label = "Frecuencia: reuniones publicas/comunitarias")


lab_study1 <- sjlabelled::get_label(df_study1) 
df_study1 <- df_study1 %>% na.omit() %>% sjlabelled::set_label(lab_study1)

df<- dfSummary(x = dplyr::select(df_study1, ola, marchar, starts_with("barrio_"), starts_with("satisfaccion_"), educ, quintil1, ess, edad, sexo, pos_id),
               plain.ascii = FALSE,
               style = "grid",
               tmp.img.dir = "/tmp",
               graph.magnif = TRUE,
               headings = F,  # encabezado
               varnumbers = F, # num variable
               labels.col = T, # etiquetas
               na.col = F,    # missing
               graph.col = T, # plot
               valid.col = F, # n valido
               col.widths = c(700,10,10,10)
               )
df$Variable <- NULL # delete variable column
summarytools::view(df,
     file = here::here("output/tables/desc02.html")
     ) # Ver tabla en un archivo HTML
webshot::webshot(url =here::here("output/tables/desc02.html"),
        file =here::here("output/tables/desc02.png")) # Remover archivo HTML
# print(df, method = "render")
```

```{r}
load(file = here::here("input/data-proc/df_study1_long_t7.RData"))
df_study1 <- df_study1_long_t7 %>% dplyr::select(idencuesta, ola, ponderador_long_total, segmento, estrato,
                                                marchar,
         barrio_ideal,
         barrio_integrado,
         barrio_identifico,
         barrio_partedemi,
         barrio_amigos,
         barrio_sociable,
         barrio_cordial,
         barrio_colaboradora,
         satisfaccion_proximidadtrabajo,
         satisfaccion_proximidadcolegios,
         satisfaccion_proximidadcomercio,
         satisfaccion_seguridad,
         jjvv_prom,
         org_religiosa_prom,
         org_caridad_prom,
         org_deportiva_prom,
                                                educ, quintil1, ess, edad, sexo, pos_id) %>% na.omit()
df_study1 <- df_study1[complete.cases(df_study1$segmento), ]

df_study1$marchar <- factor(df_study1$marchar,
         labels = c("Nunca", "Casi nunca", "A veces", "Frecuentemente", "Muy 
frecuentemente"))

elsoc_diseno <- svydesign(ids = ~segmento, #muestreo por conglomerado a nivel de manzanas (segmento)
                          strata = ~estrato, #muestreo estratificado a nivel ciudad (estato)
                          weights = ~ponderador_long_total, #ponderador longitudinal
                          nest = TRUE,
                          data = df_study1)


#Paso 1
datos.marchar <- data.frame((svytable(~marchar + ola + idencuesta, elsoc_diseno, round = F))) %>% dplyr::filter(Freq>0)  %>% group_by(ola) %>% mutate(porcentaje=Freq/sum(Freq)) %>% na.omit()

#Paso 2
etiquetas.marchar <- data.frame((svytable(~marchar + ola, elsoc_diseno, round = F))) %>% group_by(ola) %>% mutate(porcentaje=Freq/sum(Freq)) %>% na.omit() %>% 
  mutate(idencuesta = 1)
datos.marchar$marchar <- factor(datos.marchar$marchar, levels = rev(levels(datos.marchar$marchar)))
etiquetas.marchar$marchar <- factor(etiquetas.marchar$marchar, levels = rev(levels(etiquetas.marchar$marchar)))
```

```{r alluvial}
colors<- c("#f1eef6ff","#bdc9e1ff","#b3b3b3ff","#74a9cfff","#0570b0ff")
alluvial_marchar <- ggplot(datos.marchar, aes(x = ola, fill = marchar, stratum = marchar,
                             alluvium = idencuesta, y = porcentaje))+
    ggalluvial::geom_flow(alpha = .66) + 
    ggalluvial::geom_stratum(linetype = 0) +
    scale_y_continuous(labels = scales::percent) + 
    ylab(label = NULL) +
    xlab(label = NULL) + 
    theme(#legend.position = 'top',
          legend.title = element_blank(),
          plot.caption = element_text(hjust = 1)) +
    scale_fill_manual(values = colors) +
    geom_shadowtext(data = etiquetas.marchar, 
              aes(label = ifelse(porcentaje > 0 , scales::percent(porcentaje, accuracy = .1),"")),
              position = position_stack(vjust = .5),
              show.legend = FALSE,
              size = 4,
              color = rep('white'),
              bg.colour='grey30')+
  scale_x_discrete(labels = c(2016, 2017, 2018, 2019, 2022, 2023))

alluvial_marchar

ggsave(alluvial_marchar, file="output/graphs/alluvial_dep.png", height = 8, width = 7)
```

```{r bivariate-plots}
df_study1 <- df_study1_long_t7 %>% dplyr::select(idencuesta, ola, ponderador_long_total, segmento, estrato,
                                               marchar, starts_with("barrio_"),
                                                educ, quintil1, ess, edad, sexo, pos_id) %>% na.omit()
cohesion_plot <- df_study1 %>% 
  select(starts_with("barrio_"), starts_with("satisfaccion_")) %>% 
  sjPlot::plot_stackfrq(weight.by = df_study1$ponderador_long_total,
                        expand.grid =TRUE,
                      geom.colors = "PuBu",
                      show.total = FALSE, 
         #             vjust=rep(c("bottom", "top", "bottom", "top", "bottom"),4),
                      legend.labels = c("Muy en desacuerdo", "Desacuerdo", "Ni de acuerdo
Ni desacuerdo", "De acuerdo", "Muy de acuerdo"),
#                       axis.labels = c("Merit effort", "Merit talent", "Inequality 
# perception", "Justification 
# of 
# educational inequality")
                      ) +
  theme_light(base_size = 16) +
  theme(legend.position="bottom", text = element_text(size = 14),
        plot.caption = element_text(hjust = 1))
cohesion_plot

ggsave(cohesion_plot, file="output/graphs/cohesion_plot.png", height = 6, width = 10)
```

```{r bivariate-plots2}
df_study1 <- df_study1_long_t7 %>% dplyr::select(idencuesta, ola, ponderador_long_total, segmento, estrato,
                                               marchar, starts_with("satisfaccion_"),
                                                educ, quintil1, ess, edad, sexo, pos_id) %>% na.omit()
satisfaccion_plot <- df_study1 %>% 
  select(starts_with("barrio_"), starts_with("satisfaccion_")) %>% 
  sjPlot::plot_stackfrq(weight.by = df_study1$ponderador_long_total,
                        expand.grid =TRUE,
                      geom.colors = "PuBu",
                      show.total = FALSE, 
         #             vjust=rep(c("bottom", "top", "bottom", "top", "bottom"),4),
                      legend.labels = c("Muy en desacuerdo", "Desacuerdo", "Ni de acuerdo
Ni desacuerdo", "De acuerdo", "Muy de acuerdp"),
#                       axis.labels = c("Merit effort", "Merit talent", "Inequality 
# perception", "Justification 
# of 
# educational inequality")
                      ) +
  theme_light(base_size = 16) +
  theme(legend.position="bottom", text = element_text(size = 14),
        plot.caption = element_text(hjust = 1))
satisfaccion_plot

ggsave(satisfaccion_plot, file="output/graphs/satisfaccion_plot.png", height = 6, width = 10)
```


# Regresión


```{r marchar}
rm(list=ls())
#if (!require("pacman")) install.packages("pacman") # instalar pacman
                            # cargar librerias
pacman::p_load(dplyr,       # Manipulacion de datos
               knitr,
               kableExtra,
               summarytools,
               lme4,
               texreg
               )
load(file = here::here("input/data-proc/df_study1_long_t7.RData"))

# generate analytical sample
df_study1 <- 
  df_study1_long_t7 %>%
  select(idencuesta,ola,comuna, region_cod, ponderador_long_total, marchar, part_com, satisfaccion_seguridad, starts_with("barrio_"), jjvv_prom, org_religiosa_prom, org_caridad_prom, org_deportiva_prom, educ, quintil1, ess, edad, sexo, pos_id) %>% 
  na.omit() %>% 
  mutate(ola_num=as.numeric(ola),
         ola_2 = ola_num^2,
         ola=as.factor(ola),
         sexo=as.factor(sexo)
         )
# Hipotesis
h1 <- "ola"
h2 <- "ola_num"
h3 <- "ola_2"
h4 <- "ola_num+ola_2"

marchar.null <- lmer(formula(paste0("marchar~","1 + (1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar1 <- lmer(formula(paste0("marchar~",h1,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar2 <- lmer(formula(paste0("marchar~",h2,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar3 <- lmer(formula(paste0("marchar~",h3,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar4 <- lmer(formula(paste0("marchar~",h4,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

performance::icc(marchar.null)
```

```{r marchar-reg, results='asis'}
texreg::htmlreg(list(marchar.null, marchar1, marchar2, marchar3, marchar4),
                custom.note = "*** p < 0.001; ** p < 0.01; * p < 0.05")
```


```{r marchar2-cor}
sjmisc::frq(df_study1$region_cod)
df_study1 <- df_study1 %>% mutate(rm= case_when(region_cod==13~"RM",
                                                TRUE~"Otra región"))
sjmisc::frq(df_study1$rm)

df_study1 %>% filter(ola==1) %>% select(barrio_ideal:barrio_colaboradora) %>% sjPlot::tab_fa(method = "ml", rotation = "promax", show.comm = TRUE, title = "Análisis factorial exploratorio")

M <-df_study1 %>% filter(ola==1) %>% select(barrio_ideal:barrio_colaboradora) %>% cor(use = "complete.obs") # Usar solo casos con observaciones completas
diag(M) = NA # Elimina la diagonal (correlaciones absolutas de cada variable consigmo misma)
# Set Row names of the matrix
rownames(M) <- c("A. Este es el barrio ideal para mi",
                     "B. Me siento integrado/a en este barrio",
                     "C. Me identifico con la gente de este barrio",
                     "D. Este barrio es parte de mi",
                     "E. En este barrio es fácil hacer amigos",
                     "F. La gente en este barrio es sociable",
                     "G. La gente en este barrio es cordial",
                     "H. La gente en este barrio es colaboradora")
#set Column names of the matrix
colnames(M) <-c("(A)", "(B)","(C)","(D)","(E)","(F)","(G)", "(H)")
```


```{r marchar2-cor2}
#Plot the matrix using corrplot
corrplot::corrplot(M,
  method = "color", # Cambia los círculos por color completo de cada cuadrante
  addCoef.col = "#000390", # Color de los coeficientes
  type = "upper", # Deja solo las correlaciones de arriba
  tl.col = "black", # COlor letras, rojo por defecto
  na.label = "-")
```


```{r marchar2-cor3}
df_study1 <- df_study1 %>% 
  rowwise() %>%
  mutate(attachment = mean(c(barrio_ideal, barrio_integrado, barrio_identifico, barrio_partedemi, na.rm=TRUE)),
         sociabilidad = mean(c(barrio_amigos, barrio_sociable, barrio_cordial, barrio_colaboradora, na.rm=TRUE))) %>% 
  ungroup()

M2 <-df_study1 %>% filter(ola==1) %>% select(attachment, sociabilidad) %>% cor(use = "complete.obs") # Usar solo casos con observaciones completas
diag(M2) = NA # Elimina la diagonal (correlaciones absolutas de cada variable consigmo misma)
# Set Row names of the matrix
rownames(M2) <- c("A. Attachment",
                     "B. Sociabilidad")
#set Column names of the matrix
colnames(M2) <-c("(A)", "(B)")
```


```{r marchar2-cor4}
#Plot the matrix using corrplot
corrplot::corrplot(M2,
  method = "color", # Cambia los círculos por color completo de cada cuadrante
  addCoef.col = "#000390", # Color de los coeficientes
  type = "upper", # Deja solo las correlaciones de arriba
  tl.col = "black", # COlor letras, rojo por defecto
  na.label = "-")

```

```{r marchar2}
h5 <- "ola+edad"
h6 <- "ola+edad+rm"
h7 <- "ola+edad+rm+satisfaccion_seguridad"
h8 <- "ola+edad+rm+attachment"
h9 <- "ola+edad+rm+sociabilidad"
h10 <- "ola+edad+rm+part_com"
h11 <- "ola+edad+rm+jjvv_prom"
h12 <- "ola+edad+rm+org_religiosa_prom"
h13 <- "ola+edad+rm+org_caridad_prom"
h14 <- "ola+edad+rm+org_deportiva_prom"
h15 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom"
h16 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h17 <- "ola*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h17.2 <- "edad*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h18 <- "ola*satisfaccion_seguridad+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h18.2 <- "ola+satisfaccion_seguridad*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h19 <- "edad*satisfaccion_seguridad*ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h20 <- "ola+attachment*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h21 <- "edad*attachment*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h22 <- "ola+sociabilidad*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h23 <- "edad*sociabilidad*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h24 <- "ola+part_com*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h25 <- "edad*part_com*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h26 <- "ola+jjvv_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h27 <- "edad*jjvv_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h28 <- "ola+org_religiosa_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h29 <- "edad*org_religiosa_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h30 <- "ola+org_caridad_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h31 <- "edad*org_caridad_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h32 <- "ola+org_deportiva_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h33 <- "edad*org_deportiva_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"


# A. Education distributive justice

marchar5 <- lmer(formula(paste0("marchar~",h5,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar6 <- lmer(formula(paste0("marchar~",h6,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar7 <- lmer(formula(paste0("marchar~",h7,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar8 <- lmer(formula(paste0("marchar~",h8,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar9 <- lmer(formula(paste0("marchar~",h9,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar10 <- lmer(formula(paste0("marchar~",h10,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar11 <- lmer(formula(paste0("marchar~",h11,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar12 <- lmer(formula(paste0("marchar~",h12,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar13 <- lmer(formula(paste0("marchar~",h13,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar14 <- lmer(formula(paste0("marchar~",h14,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar15 <- lmer(formula(paste0("marchar~",h15,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar16 <- lmer(formula(paste0("marchar~",h16,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar17.2 <- lmer(formula(paste0("marchar~",h17.2,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar17 <- lmer(formula(paste0("marchar~",h17,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar18.2 <- lmer(formula(paste0("marchar~",h18.2,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar18 <- lmer(formula(paste0("marchar~",h18,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar19 <- lmer(formula(paste0("marchar~",h19,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar20 <- lmer(formula(paste0("marchar~",h20,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar21 <- lmer(formula(paste0("marchar~",h21,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar22 <- lmer(formula(paste0("marchar~",h22,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar23 <- lmer(formula(paste0("marchar~",h23,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar24 <- lmer(formula(paste0("marchar~",h24,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar25 <- lmer(formula(paste0("marchar~",h25,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar26 <- lmer(formula(paste0("marchar~",h26,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar27 <- lmer(formula(paste0("marchar~",h27,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar28 <- lmer(formula(paste0("marchar~",h28,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)

marchar29 <- lmer(formula(paste0("marchar~",h29,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar30 <- lmer(formula(paste0("marchar~",h30,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar31 <- lmer(formula(paste0("marchar~",h31,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar32 <- lmer(formula(paste0("marchar~",h32,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar33 <- lmer(formula(paste0("marchar~",h33,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
```

## Modelos de regresión (con organizaciones de caridad)

```{r marchar2-reg, results='asis'}
#omit <- "(pos_id)|(edad)|(cuestion_mig)|(nation_t1)|(quintil1QNA)|(Intercept)"
texreg::htmlreg(list(marchar1, marchar5, marchar6, marchar7, marchar8, marchar9, marchar10, marchar11, marchar12, marchar13, marchar14, marchar15, marchar16),
                custom.note = "*** p < 0.001; ** p < 0.01; * p < 0.05", 
#                custom.coef.names = c("Intercepto",
#                                      "ola 2017 <br> <i>(Ref. ola 2016)</i>",
#                                      "ola 2018",
#                                      "ola 2019",
##                                      "ola 2022",
#                                      "Merito: esfuerzo",
#                                      "Merito: Inteligencia",
##                                      "Goahead: Familia",
#                                      "Goahead: educacion",
#                                      "Goahead: ambicion",
#                                      "Goahead: trabajo duro",
#                                      "Ed. media <br> <i>(Ref. Ed. básica)</i>",
#                                      "Ed. técnica",
#                                      "Ed. universitaria",
#                                      "Quintil 2 <br> <i>(Ref. Quintil 1)</i>",
#                                      "Quintil 3",
#                                      "Quintil 4",
#                                      "Quintil 5",
#                                      "Quintil NA",
#                                      "Estatus social subjetivo",
#                                      "Pos. política: Centro <br> <i>(Ref. izquierda)</i>",
#                                      "Pos. política: Derecha",
#                                      "Pos. política: No se identifica",
#                                      "Edad 30-49 <br> <i>(Ref. 18-29)</i>",
#                                      "Edad 50-64",
#                                      "Edad 65 o más",
#                                      "Mujer <br> <i>(Ref. Hombre)</i>",
#                                      "Gini comunal",
#                                      "Cambio gini 2022-2017",
#                                      "Proporción educ universitaria",
#                                      "Proporción matrícula privada",
#                                      "Total matrícula privada",
#                                      "Varianza Simce comuna",
#                                      "Promedio Simce comuna"
                                    
                              )
          #omit.coef = omit

```

## Modelos de regresión (sin organizaciones de caridad)

```{r marchar3}
h5 <- "ola+edad"
h6 <- "ola+edad+rm"
h7 <- "ola+edad+rm+satisfaccion_seguridad"
h8 <- "ola+edad+rm+attachment"
h9 <- "ola+edad+rm+sociabilidad"
h10 <- "ola+edad+rm+part_com"
h11 <- "ola+edad+rm+jjvv_prom"
h12 <- "ola+edad+rm+org_religiosa_prom"
h13 <- "ola+edad+rm+org_deportiva_prom"
h14 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom"
h15 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
```


```{r marchar3-interacts}
h16 <- "ola*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h17 <- "satisfaccion_seguridad*ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h18 <- "ola+satisfaccion_seguridad*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h19 <- "edad*satisfaccion_seguridad*ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h20 <- "ola+attachment*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h21 <- "edad*attachment*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h22 <- "ola+sociabilidad*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h23 <- "edad*sociabilidad*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h24 <- "ola+part_com*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h25 <- "edad*part_com*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h26 <- "ola+jjvv_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h27 <- "edad*jjvv_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h28 <- "ola+org_religiosa_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h29 <- "edad*org_religiosa_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h30 <- "ola+org_deportiva_prom*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"
h31 <- "edad*org_deportiva_prom*ola+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_deportiva_prom+educ+quintil1+sexo+ess+pos_id"


# A. Education distributive justice

marchar5 <- lmer(formula(paste0("marchar~",h5,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar6 <- lmer(formula(paste0("marchar~",h6,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar7 <- lmer(formula(paste0("marchar~",h7,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar8 <- lmer(formula(paste0("marchar~",h8,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar9 <- lmer(formula(paste0("marchar~",h9,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar10 <- lmer(formula(paste0("marchar~",h10,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar11 <- lmer(formula(paste0("marchar~",h11,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar12 <- lmer(formula(paste0("marchar~",h12,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar13 <- lmer(formula(paste0("marchar~",h13,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar14 <- lmer(formula(paste0("marchar~",h14,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar15 <- lmer(formula(paste0("marchar~",h15,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar16 <- lmer(formula(paste0("marchar~",h16,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar17 <- lmer(formula(paste0("marchar~",h17,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar18 <- lmer(formula(paste0("marchar~",h18,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar19 <- lmer(formula(paste0("marchar~",h19,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar20 <- lmer(formula(paste0("marchar~",h20,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar21 <- lmer(formula(paste0("marchar~",h21,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar22 <- lmer(formula(paste0("marchar~",h22,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar23 <- lmer(formula(paste0("marchar~",h23,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar24 <- lmer(formula(paste0("marchar~",h24,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar25 <- lmer(formula(paste0("marchar~",h25,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar26 <- lmer(formula(paste0("marchar~",h26,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar27 <- lmer(formula(paste0("marchar~",h27,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar28 <- lmer(formula(paste0("marchar~",h28,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar29 <- lmer(formula(paste0("marchar~",h29,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar30 <- lmer(formula(paste0("marchar~",h30,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
marchar31 <- lmer(formula(paste0("marchar~",h31,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total)
```

```{r marchar3-reg, results='asis'}
#omit <- "(pos_id)|(edad)|(cuestion_mig)|(nation_t1)|(quintil1QNA)|(Intercept)"
texreg::htmlreg(list(marchar1, marchar5, marchar6, marchar7, marchar8, marchar9, marchar10, marchar11, marchar12, marchar13, marchar14, marchar15),
                custom.note = "*** p < 0.001; ** p < 0.01; * p < 0.05", 
#                custom.coef.names = c("Intercepto",
#                                      "ola 2017 <br> <i>(Ref. ola 2016)</i>",
#                                      "ola 2018",
#                                      "ola 2019",
##                                      "ola 2022",
#                                      "Merito: esfuerzo",
#                                      "Merito: Inteligencia",
##                                      "Goahead: Familia",
#                                      "Goahead: educacion",
#                                      "Goahead: ambicion",
#                                      "Goahead: trabajo duro",
#                                      "Ed. media <br> <i>(Ref. Ed. básica)</i>",
#                                      "Ed. técnica",
#                                      "Ed. universitaria",
#                                      "Quintil 2 <br> <i>(Ref. Quintil 1)</i>",
#                                      "Quintil 3",
#                                      "Quintil 4",
#                                      "Quintil 5",
#                                      "Quintil NA",
#                                      "Estatus social subjetivo",
#                                      "Pos. política: Centro <br> <i>(Ref. izquierda)</i>",
#                                      "Pos. política: Derecha",
#                                      "Pos. política: No se identifica",
#                                      "Edad 30-49 <br> <i>(Ref. 18-29)</i>",
#                                      "Edad 50-64",
#                                      "Edad 65 o más",
#                                      "Mujer <br> <i>(Ref. Hombre)</i>",
#                                      "Gini comunal",
#                                      "Cambio gini 2022-2017",
#                                      "Proporción educ universitaria",
#                                      "Proporción matrícula privada",
#                                      "Total matrícula privada",
#                                      "Varianza Simce comuna",
#                                      "Promedio Simce comuna"
                                    
                              )
          #omit.coef = omit

```

### interacciones

```{r marchar-interac, results='asis'}
#omit <- "(pos_id)|(edad)|(cuestion_mig)|(nation_t1)|(quintil1QNA)|(Intercept)"
texreg::htmlreg(list(marchar16, marchar17, marchar18, marchar19, marchar20,
                     marchar21,marchar22,marchar23,marchar24,marchar25,
                     marchar26,marchar27,marchar28, marchar29, marchar30, marchar31),
                custom.note = "*** p < 0.001; ** p < 0.01; * p < 0.05", 
#                custom.coef.names = c("Intercepto",
#                                      "ola 2017 <br> <i>(Ref. ola 2016)</i>",
#                                      "ola 2018",
#                                      "ola 2019",
##                                      "ola 2022",
#                                      "Merito: esfuerzo",
#                                      "Merito: Inteligencia",
##                                      "Goahead: Familia",
#                                      "Goahead: educacion",
#                                      "Goahead: ambicion",
#                                      "Goahead: trabajo duro",
#                                      "Ed. media <br> <i>(Ref. Ed. básica)</i>",
#                                      "Ed. técnica",
#                                      "Ed. universitaria",
#                                      "Quintil 2 <br> <i>(Ref. Quintil 1)</i>",
#                                      "Quintil 3",
#                                      "Quintil 4",
#                                      "Quintil 5",
#                                      "Quintil NA",
#                                      "Estatus social subjetivo",
#                                      "Pos. política: Centro <br> <i>(Ref. izquierda)</i>",
#                                      "Pos. política: Derecha",
#                                      "Pos. política: No se identifica",
#                                      "Edad 30-49 <br> <i>(Ref. 18-29)</i>",
#                                      "Edad 50-64",
#                                      "Edad 65 o más",
#                                      "Mujer <br> <i>(Ref. Hombre)</i>",
#                                      "Gini comunal",
#                                      "Cambio gini 2022-2017",
#                                      "Proporción educ universitaria",
#                                      "Proporción matrícula privada",
#                                      "Total matrícula privada",
#                                      "Varianza Simce comuna",
#                                      "Promedio Simce comuna"
                                    
                              )
          #omit.coef = omit

```

ola x edad

```{r}
sim_slopes(marchar16, pred = edad, modx = ola, interval = TRUE)
```

```{r}
interact_plot(marchar16, pred = edad, modx = ola, interval = TRUE)
```


satisfaccion seguridad x edad

```{r}
sim_slopes(marchar18, pred = satisfaccion_seguridad, modx = edad, jnplot = TRUE)
```

ola x satisfaccion_seguridad x edad

```{r}
sim_slopes(marchar19, pred = satisfaccion_seguridad, modx = edad, mod2 = ola, jnplot = TRUE)
```

ola x attachment

```{r}
sim_slopes(marchar20, pred = attachment, modx = edad, jnplot = TRUE)
```

ola x attachment x edad

```{r}
sim_slopes(marchar21, pred = attachment, modx = edad, mod2 = ola, jnplot = TRUE)
```

ola x sociabilidad

```{r}
sim_slopes(marchar22, pred = sociabilidad, modx = edad, jnplot = TRUE)
```

ola x sociabilidad x edad

```{r}
sim_slopes(marchar23, pred = sociabilidad, modx = edad, mod2 = ola, jnplot = TRUE)
```

ola x part_com

```{r}
sim_slopes(marchar24, pred = part_com, modx = edad, jnplot = TRUE)
```

ola x part_com x edad

```{r}
sim_slopes(marchar25, pred = part_com, modx = edad, mod2 = ola, jnplot = TRUE)
```

ola x juntas de vecinos

```{r}
sim_slopes(marchar26, pred = jjvv_prom, modx = edad, jnplot = TRUE)
```

ola x juntas de vecinos x edad

```{r}
sim_slopes(marchar27, pred = jjvv_prom, modx = edad, mod2 = ola, jnplot = TRUE)
```

ola x org religiosas

```{r}
sim_slopes(marchar28, pred = org_religiosa_prom, modx = edad, jnplot = TRUE)
```

ola x org religiosas x edad

```{r}
sim_slopes(marchar29, pred = org_religiosa_prom, modx = edad, mod2 = ola, jnplot = TRUE)
```

ola x org deportivas

```{r}
sim_slopes(marchar30, pred = org_deportiva_prom, modx = edad, jnplot = TRUE)
```

ola x org deportivas x edad

```{r}
sim_slopes(marchar31, pred = org_deportiva_prom, modx = edad, mod2 = ola, jnplot = TRUE)
```
