==> production/03-resultados-v2.qmd <==
---
title: "03-resultados"
format: 
  html:
    toc: true
editor: source
lang: es
---

```{r setup, echo=FALSE}
knitr::opts_chunk$set(
	echo = FALSE
)
options(scipen=9999) # desactivar notacion cientifica
remove(list = ls()) #limpieza del entorno de trabajo
```

# Participación electoral y cohesión barrial

*Reporte de resultados*

## Participación electoral

Variables:

- Participación elecciones 2013 (ola 1 y 2)

- Participación elecciones 2017 (ola 3 y 4)

- Participación en plebiscito nueva constitución (ola 5)

- Participación elecciones 2021 primera vuelta (ola 6)

- Participación elecciones 2021 segunda vuelta (ola 7)

```{r }
pacman::p_load(dplyr,       # Manipulacion de datos
               knitr,
               kableExtra,
               summarytools,
               ggplot2,
               sjPlot,
               tidyverse,
               ggalluvial,
               survey,
               shadowtext,
               srvyr,
               interactions,
               lme4
               )
```

```{r}
load(file = here::here("input/data-proc/df_study1_long_t7.RData"))
df_study1 <- df_study1_long_t7 %>% dplyr::select(idencuesta, ola, ponderador_long_total, segmento, estrato,
                                                part_electoral,
         barrio_ideal,
         barrio_integrado,
         barrio_identifico,
         barrio_partedemi,
         barrio_amigos,
         barrio_sociable,
         barrio_cordial,
         barrio_colaboradora,
         satisfaccion_seguridad,
         jjvv_prom,
         org_religiosa_prom,
         org_caridad_prom,
         org_deportiva_prom,
         educ, quintil1, edad, sexo, pos_id) %>% na.omit()

df_study1$part_electoral <- factor(df_study1$part_electoral, levels = c(1,2),
         labels = c("Si", "No"))
df_study1$part_electoral <- car::recode(df_study1$part_electoral, c("1=2; 2=1"))
df_study1 <- df_study1 %>%
  dplyr::mutate(part_electoral = ifelse(part_electoral == "Si", 1, 0))

df_study1$part_electoral <- factor(df_study1$part_electoral,
         labels = c("No", "Si"))

df_study1 <- df_study1[complete.cases(df_study1$segmento), ]



elsoc_diseno <- svydesign(ids = ~segmento, #muestreo por conglomerado a nivel de manzanas (segmento)
                          strata = ~estrato, #muestreo estratificado a nivel ciudad (estato)
                          weights = ~ponderador_long_total, #ponderador longitudinal
                          nest = TRUE,
                          data = df_study1)


#Paso 1
datos.marchar <- data.frame((svytable(~part_electoral + ola + idencuesta, elsoc_diseno, round = F))) %>% dplyr::filter(Freq>0)  %>% group_by(ola) %>% mutate(porcentaje=Freq/sum(Freq)) %>% na.omit()

#Paso 2
etiquetas.marchar <- data.frame((svytable(~part_electoral + ola, elsoc_diseno, round = F))) %>% group_by(ola) %>% mutate(porcentaje=Freq/sum(Freq)) %>% na.omit() %>% 
  mutate(idencuesta = 1)
#datos.marchar$part_electoral <- factor(datos.marchar$part_electoral, levels = rev(levels(datos.marchar$part_electoral)))
#etiquetas.marchar$part_electoral <- factor(etiquetas.marchar$part_electoral, levels = rev(levels(etiquetas.marchar$part_electoral)))
```

## Cambios en la participación en manifestaciones entre 2016 y 2023

```{r alluvial}
colors<- c("#f1eef6ff","#0570b0ff")
alluvial_marchar <- ggplot(datos.marchar, aes(x = ola, fill = part_electoral, stratum = part_electoral,
                             alluvium = idencuesta, y = porcentaje))+
    ggalluvial::geom_flow(alpha = .66) + 
    ggalluvial::geom_stratum(linetype = 0) +
    scale_y_continuous(labels = scales::percent) + 
    ylab(label = NULL) +
    xlab(label = NULL) + 
    theme(#legend.position = 'top',
          legend.title = element_blank(),
          plot.caption = element_text(hjust = 1)) +
    scale_fill_manual(values = colors) +
    geom_shadowtext(data = etiquetas.marchar, 
              aes(label = ifelse(porcentaje > 0 , scales::percent(porcentaje, accuracy = .1),"")),
              position = position_stack(vjust = .5),
              show.legend = FALSE,
              size = 4,
              color = rep('white'),
              bg.colour='grey30')+
  scale_x_discrete(labels = c("Presidenciales
2013", "Presidenciales
2013", "Presidenciales
2017", "Presidenciales
2017", "Plebiscito 
nueva constitución","Primera vuelta
2021", "Segunda vuelta
2021"))
```


```{r alluvial-2}
#| label: fig-alluvial
#| fig-cap: "Cambios en la participación en manifestaciones entre 2016 y 2023"

alluvial_marchar
```

La @fig-alluvial muestra un porcentaje de participación electoral entre el 65% y 80%. Sin embargo, es notorio el aumento en la cantidad de participantes en las últimas dos elecciones (primera y segunda vuelta presidenciales 2021) en comparación con los años anteriores.

## Modelos de regresión multinivel

```{r, warning=FALSE, cache=TRUE}
rm(list=ls())
#if (!require("pacman")) install.packages("pacman") # instalar pacman
                            # cargar librerias
pacman::p_load(dplyr,       # Manipulacion de datos
               knitr,
               kableExtra,
               summarytools,
               lme4,
               texreg
               )
load(file = here::here("input/data-proc/df_study1_long_t7.RData"))

# generate analytical sample
df_study1 <- 
  df_study1_long_t7 %>%
  select(idencuesta,ola,comuna, region_cod, ponderador_long_total, part_electoral, part_com, satisfaccion_seguridad, starts_with("barrio_"), jjvv_prom, org_religiosa_prom, org_caridad_prom, org_deportiva_prom, educ, quintil1, ess, edad, sexo, pos_id) %>% 
  na.omit() %>% 
  mutate(ola_num=as.numeric(ola),
         ola_2 = ola_num^2,
         ola=as.factor(ola),
         sexo=as.factor(sexo)
         )

df_study1$part_electoral <- factor(df_study1$part_electoral, levels = c(1,2),
         labels = c("Si", "No"))
df_study1$part_electoral <- car::recode(df_study1$part_electoral, c("1=2; 2=1"))
df_study1 <- df_study1 %>%
  dplyr::mutate(part_electoral = ifelse(part_electoral == "Si", 1, 0))

# df_study1$part_electoral <- factor(df_study1$part_electoral,
#          labels = c("No", "Si"))

df_study1 <- df_study1 %>% mutate(rm= case_when(region_cod==13~"RM",
                                                TRUE~"Otra región"))

df_study1 <- df_study1 %>% 
  rowwise() %>%
  mutate(attachment = mean(c(barrio_ideal, barrio_integrado, barrio_identifico, barrio_partedemi, na.rm=TRUE)),
         sociabilidad = mean(c(barrio_amigos, barrio_sociable, barrio_cordial, barrio_colaboradora, na.rm=TRUE))) %>% 
  ungroup()

h5 <- "ola+edad"
h6 <- "ola+edad+rm"
h7 <- "ola+edad+rm+satisfaccion_seguridad"
h8 <- "ola+edad+rm+attachment"
h9 <- "ola+edad+rm+sociabilidad"
h10 <- "ola+edad+rm+part_com"
h11 <- "ola+edad+rm+jjvv_prom"
h12 <- "ola+edad+rm+org_religiosa_prom"
h13 <- "ola+edad+rm+org_caridad_prom"
h14 <- "ola+edad+rm+org_deportiva_prom"
h15 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom"
h16 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id"
h17 <- "ola*edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id"
h18 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*satisfaccion_seguridad"
h19 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*attachment"
h20 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*sociabilidad"
h21 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*part_com"
h22 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*jjvv_prom"
h23 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*org_religiosa_prom"
h24 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*org_caridad_prom"
h25 <- "ola+edad+rm+satisfaccion_seguridad+attachment+sociabilidad+part_com+jjvv_prom+org_religiosa_prom+org_caridad_prom+org_deportiva_prom+educ+quintil1+sexo+pos_id+edad*org_deportiva_prom"

marchar5 <- glmer(formula(paste0("part_electoral~",h5,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar6 <- glmer(formula(paste0("part_electoral~",h6,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar7 <- glmer(formula(paste0("part_electoral~",h7,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar8 <- glmer(formula(paste0("part_electoral~",h8,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar9 <- glmer(formula(paste0("part_electoral~",h9,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar10 <- glmer(formula(paste0("part_electoral~",h10,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar11 <- glmer(formula(paste0("part_electoral~",h11,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar12 <- glmer(formula(paste0("part_electoral~",h12,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar13 <- glmer(formula(paste0("part_electoral~",h13,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar14 <- glmer(formula(paste0("part_electoral~",h14,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar15 <- glmer(formula(paste0("part_electoral~",h15,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar16 <- glmer(formula(paste0("part_electoral~",h16,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar17 <- glmer(formula(paste0("part_electoral~",h17,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar18 <- glmer(formula(paste0("part_electoral~",h18,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar19 <- glmer(formula(paste0("part_electoral~",h19,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar20 <- glmer(formula(paste0("part_electoral~",h20,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar21 <- glmer(formula(paste0("part_electoral~",h21,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar22 <- glmer(formula(paste0("part_electoral~",h22,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar23 <- glmer(formula(paste0("part_electoral~",h23,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar24 <- glmer(formula(paste0("part_electoral~",h24,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar25 <- glmer(formula(paste0("part_electoral~",h25,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
```

```{r marchar2-reg, results='asis'}
#| label: tbl-reg
#| tbl-cap: "Modelos de regresión multinivel"


#omit <- "(pos_id)|(edad)|(cuestion_mig)|(nation_t1)|(quintil1QNA)|(Intercept)"
texreg::knitreg(list(marchar5, marchar6, marchar7, marchar8, marchar9, marchar10, marchar11, marchar12, marchar13, marchar14, marchar15, marchar16),
                custom.note = "*** p < 0.001; ** p < 0.01; * p < 0.05", 
#                custom.coef.names = c("Intercepto",
#                                      "ola 2017 <br> <i>(Ref. ola 2016)</i>",
#                                      "ola 2018",
#                                      "ola 2019",
##                                      "ola 2022",
#                                      "Merito: esfuerzo",
#                                      "Merito: Inteligencia",
##                                      "Goahead: Familia",
#                                      "Goahead: educacion",
#                                      "Goahead: ambicion",
#                                      "Goahead: trabajo duro",
#                                      "Ed. media <br> <i>(Ref. Ed. básica)</i>",
#                                      "Ed. técnica",
#                                      "Ed. universitaria",
#                                      "Quintil 2 <br> <i>(Ref. Quintil 1)</i>",
#                                      "Quintil 3",
#                                      "Quintil 4",
#                                      "Quintil 5",
#                                      "Quintil NA",
#                                      "Estatus social subjetivo",
#                                      "Pos. política: Centro <br> <i>(Ref. izquierda)</i>",
#                                      "Pos. política: Derecha",
#                                      "Pos. política: No se identifica",
#                                      "Edad 30-49 <br> <i>(Ref. 18-29)</i>",
#                                      "Edad 50-64",
#                                      "Edad 65 o más",
#                                      "Mujer <br> <i>(Ref. Hombre)</i>",
#                                      "Gini comunal",
#                                      "Cambio gini 2022-2017",
#                                      "Proporción educ universitaria",
#                                      "Proporción matrícula privada",
#                                      "Total matrícula privada",
#                                      "Varianza Simce comuna",
#                                      "Promedio Simce comuna"
                                    
                              )
          #omit.coef = omit

```

La @tbl-reg muestra 12 modelos de regresión logística multinivel utilizando la participación electoral como variable dependiente.

- El modelo 1 incluye los años de la encuesta como variable independiente, donde se puede observar que entre 2017 y 2019 no hay diferencias en comparación con el 2019. Sin embargo, existe una mayor participación electoral en las tres últimas elecciones (plebiscito nueva constitución y primera y segunda vuelta presidencial 2021) en comparación con el 2016. Todos estos efectos son significativos y se mantienen estables al controlar por el resto de las variables.

- El modelo 1 también incluye la edad como variable independiente. Esta variable tiene un efecto positivo que indica que, a mayor edad, en promedio la participación en elecciones es mayor. Efecto que es significativo y que se mantiene estable al controlar por el resto de las variables.

- El modelo 2 incluye pertenecer a la región metropolitana, usando como referencia ser de otra región. Sin embargo, el efecto no es significativo.

- El modelo 3 incluye la satisfacción con la seguridad en el barrio como variable independiente. Esta variable presenta un efecto negativo, que indica que los encuestados que están más satisfechos con la seguridad en su barrio participan en elecciones, en promedio, en menor medida.

- El modelo 4 agrega el *attachment* o apego al barrio, que tiene un positivo pero no significativo. Sin embargo, al controlar por el resto de variables este efecto sí se vuelve significativo. Este efecto es positivo, lo que indica que la gente con mayor attachment participa más en elecciones.

- El modelo 5 agrega la sociabilidad (ej. en este barrio es fácil hacer amigos). Este efecto es negativo pero no significativo. Sin embargo, al controlar por el resto de las variables este efecto se vuelve significativo, lo que indica que la sociabilidad se asocia con una menor participación en elecciones. 

- El modelo 6 agrega la participación en actividades de interés público o comunitario, que tiene un efecto positivo, lo que indica que a mayor participación comunitaria, también mayor será la participación en elecciones. Sin embargo, al controlar por las variables socioeconómicas en el último modelo este efecto deja de ser significativo.

- El modelo 7 agrega la participación en juntas de vecinos, que tiene un efecto positivo, lo que indica que a mayor participación en juntas de vecinos, también mayor será la participación en elecciones. Sin embargo, al controlar por las variables socioeconómicas en el último modelo este efecto deja de ser significativo.

- El modelo 8 agrega la participación en organizaciones religiosas, pero el efecto no es significativo.

- El modelo 9 agrega la participación en organizaciones de caridad, que tiene un efecto positivo, lo que indica que a mayor participación en organizaciones de caridad, también mayor será la participación en elecciones.

- El modelo 10 agrega la participación en organizaciones deportivas, que tiene un efecto positivo, lo que indica que a mayor participación en organizaciones deportivas, también mayor será la participación en elecciones.

- El modelo 11 agrega todas las variables anteriores a la vez. Todas las variables se mantienen estables en su sentido y, aunque varían un poco su significación, continúan siendo estadísticamente significativas.

- El modelo 12 agrega como control variables sociodemográficas como nivel educacional, quintiles de ingreso del hogar, sexo y posición política:

  * Tener educación universitaria se asocia con mayor participación en elecciones en comparación con quienes tienen educación media o menos
  * Pertenecer al quintil 7 se asocia con una mayor participación en elecciones.
  * Ser mujer no es significativo en comparación con hombres
  * Comparado con ser de izquierda, ser de centro o no identificarse con ninguno se asocia con una menor participación en elecciones.

## Tabla de interacciones

```{r, warning=FALSE}
marchar17 <- glmer(formula(paste0("part_electoral~",h17,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar18 <- glmer(formula(paste0("part_electoral~",h18,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar19 <- glmer(formula(paste0("part_electoral~",h19,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar20 <- glmer(formula(paste0("part_electoral~",h20,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar21 <- glmer(formula(paste0("part_electoral~",h21,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar22 <- glmer(formula(paste0("part_electoral~",h22,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar23 <- glmer(formula(paste0("part_electoral~",h23,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar24 <- glmer(formula(paste0("part_electoral~",h24,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
marchar25 <- glmer(formula(paste0("part_electoral~",h25,"+(1|idencuesta)")),data = df_study1, weights = ponderador_long_total, family = "binomial")
```

```{r marchar-interac, results='asis'}
#| label: tbl-interact
#| tbl-cap: "Tabla de interacciones"


#omit <- "(pos_id)|(edad)|(cuestion_mig)|(nation_t1)|(quintil1QNA)|(Intercept)"
texreg::knitreg(list(marchar16, marchar17, marchar18, marchar19, marchar20,
                     marchar21,marchar22,marchar23,marchar24,marchar25),
                custom.note = "*** p < 0.001; ** p < 0.01; * p < 0.05", 
#                custom.coef.names = c("Intercepto",
#                                      "ola 2017 <br> <i>(Ref. ola 2016)</i>",
#                                      "ola 2018",
#                                      "ola 2019",
##                                      "ola 2022",
#                                      "Merito: esfuerzo",
#                                      "Merito: Inteligencia",
##                                      "Goahead: Familia",
#                                      "Goahead: educacion",
#                                      "Goahead: ambicion",
#                                      "Goahead: trabajo duro",
#                                      "Ed. media <br> <i>(Ref. Ed. básica)</i>",
#                                      "Ed. técnica",
#                                      "Ed. universitaria",
#                                      "Quintil 2 <br> <i>(Ref. Quintil 1)</i>",
#                                      "Quintil 3",
#                                      "Quintil 4",
#                                      "Quintil 5",
#                                      "Quintil NA",
#                                      "Estatus social subjetivo",
#                                      "Pos. política: Centro <br> <i>(Ref. izquierda)</i>",
#                                      "Pos. política: Derecha",
#                                      "Pos. política: No se identifica",
#                                      "Edad 30-49 <br> <i>(Ref. 18-29)</i>",
#                                      "Edad 50-64",
#                                      "Edad 65 o más",
#                                      "Mujer <br> <i>(Ref. Hombre)</i>",
#                                      "Gini comunal",
#                                      "Cambio gini 2022-2017",
#                                      "Proporción educ universitaria",
#                                      "Proporción matrícula privada",
#                                      "Total matrícula privada",
#                                      "Varianza Simce comuna",
#                                      "Promedio Simce comuna"
                                    
                              )
          #omit.coef = omit

```

La @tbl-interact muestra las interacciones entre las variables de interes (tiempo y edad y edad con variables barriales). El modelo 1 de esta tabla es el mismo 12 de la tabla anterior.

- El modelo 2 incluye la interacción tiempo x edad. Este efecto es significativo y negativo para los años 2018 en adelante. Específicamente, en cada uno de estos años el efecto positivo de la edad se va haciendo menor. Como se puede ver en @fig-tiempo esto hace referencia a que la gente joven participa más en estos años. Destaca también que en el plebiscito por nueva constitución este efecto muestra que las personas mayores participan menos (en las demás elecciones su participación se mantiene estable)

- El modelo 3 incluye interacción edad * satisfacción de seguridad. Este efecto no es significativo, tal como se puede ver en @fig-seguridad

- El modelo 4 incluye interacción edad * attachmente. Este efecto es pequeño, pero significativo y positivo. Como se puede ver en @fig-attachment cuando hay más attachment la participación es mayor, pero el efecto en las personas mayores es aún mayor (mayor pendiente).

- El modelo 5 incluye interacción edad * sociabilidad. Este efecto es significativo y positivo, tal como se puede ver en @fig-sociabilidad cuando hay más sociabilidad las personas mayores participan más y los jóvenes participan menos.

- El modelo 6 incluye interacción edad * participación en reuniones de interés público o comunitarias. Este efecto no es significativo, tal como se puede ver en @fig-partcom

- El modelo 7 incluye interacción edad * participación en juntas de vecinos. Este efecto es significativo y positivo, tal como se puede ver en @fig-jjvv cuando hay más participación en juntas de vecinos, las personas mayores participan más y los jóvenes participan menos en elecciones.

- Los modelos 8, 9 y 10 incluyen la interacción de edad con participar en organizaciones religiosas, de caridad o deportivas. Todos estos efectos no son significativos, como se puede ver en los gráficos  @fig-religiosas, @fig-caridad y @fig-deportivas


## Gráficos de interacciones

ola x edad

```{r}
#| label: fig-tiempo


interact_plot(marchar17, pred = edad, modx = ola, jnplot = TRUE)
```

satisfaccion seguridad x edad

```{r}
#| label: fig-seguridad


interact_plot(marchar18, pred = satisfaccion_seguridad, modx = edad, jnplot = TRUE, interval = TRUE)
```

attachment x edad

```{r}
#| label: fig-attachment


interact_plot(marchar19, pred = attachment, modx = edad, jnplot = TRUE, interval = TRUE)
```

sociabilidad x edad

```{r}
#| label: fig-sociabilidad


interact_plot(marchar20, pred = sociabilidad, modx = edad, jnplot = TRUE, interval = TRUE)
```

part_com x edad

```{r}
#| label: fig-partcom


interact_plot(marchar21, pred = part_com, modx = edad, jnplot = TRUE, interval = TRUE)
```

jjvv_prom x edad

```{r}
#| label: fig-jjvv


interact_plot(marchar22, pred = jjvv_prom, modx = edad, jnplot = TRUE, interval = TRUE)
```

org_religiosa_prom x edad

```{r}
#| label: fig-religiosas


interact_plot(marchar23, pred = org_religiosa_prom, modx = edad, jnplot = TRUE, interval = TRUE)
```

org_caridad_prom x edad

```{r}
#| label: fig-caridad


interact_plot(marchar24, pred = org_caridad_prom, modx = edad, jnplot = TRUE, interval = TRUE)
```

org_deportiva_prom x edad

```{r}
#| label: fig-deportivas


interact_plot(marchar25, pred = org_deportiva_prom, modx = edad, jnplot = TRUE, interval = TRUE)
```
